+++ runtime/test/include/brt/test/common/util.h
@@ -19,11 +19,13 @@
 
 #pragma once
 
-#include "gtest/gtest.h"
+#include <stdlib.h>
+
 #include <algorithm>
 #include <iostream>
-#include <stdlib.h>
 #include <vector>
+
+#include "gtest/gtest.h"
 
 #define BRT_TEST_CHECK_STATUS(status)                                          \
   EXPECT_TRUE(status.IsOK()) << status.ErrorMessage()
@@ -90,8 +92,8 @@ template <typename T,
 // check two values same
 template <typename T> [[nodiscard]] bool ExpectEQ(T first, T second) {
   if (first != second) {
-    std::cerr << "ExpectEQ Error: first value " << first << ", second value "
-              << second << "\n";
+    std::cerr << "ExpectEQ Error: first value " << static_cast<double>(first)
+              << ", second value " << static_cast<double>(second) << "\n";
     return false;
   }
   return true;
